requirements.txt
Flask==3.1.2
numpy==1.25.2
Pillow==10.0.0
face-recognition==1.3.0
dlib==19.24.1
face_recognition_models==0.3.0
opencv-python==4.8.1.78
scipy==1.11.1
pandas==2.1.1
